# openapi.json
{"openapi":"3.1.0","info":{"title":"API Mobilité Paris – E4","description":"\nAPI REST du projet E4 – Data Mobilité Paris.\n\n## Authentification\nToutes les routes nécessitent un header **X-API-KEY**.\n\n## Données disponibles\n- 🚲 **Vélib** : stations et disponibilité temps réel\n- 🎭 **Événements** : agenda culturel parisien\n- 🛣️ **Aménagements cyclables** : pistes et voies cyclables\n","version":"1.0.0"},"paths":{"/api/velib/stations":{"get":{"tags":["🚲 Vélib"],"summary":"Liste toutes les stations Vélib","operationId":"get_stations_api_velib_stations_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/velib/stations/{station_id}":{"get":{"tags":["🚲 Vélib"],"summary":"Détail d'une station Vélib","operationId":"get_station_api_velib_stations__station_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"station_id","in":"path","required":true,"schema":{"type":"string","title":"Station Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/velib/disponibilite":{"get":{"tags":["🚲 Vélib"],"summary":"Disponibilité temps réel des vélos","operationId":"get_disponibilite_api_velib_disponibilite_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/velib/stats":{"get":{"tags":["🚲 Vélib"],"summary":"Statistiques globales Vélib","operationId":"get_stats_api_velib_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/evenements/":{"get":{"tags":["🎭 Événements"],"summary":"Liste des événements parisiens","operationId":"get_evenements_api_evenements__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"code_postal","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Ex: 75001","title":"Code Postal"},"description":"Ex: 75001"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/evenements/stats":{"get":{"tags":["🎭 Événements"],"summary":"Statistiques des événements","operationId":"get_stats_evenements_api_evenements_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/cyclable/":{"get":{"tags":["🛣️ Aménagements Cyclables"],"summary":"Liste des aménagements cyclables","operationId":"get_amenagements_api_cyclable__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"arrondissement","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Ex: 75011","title":"Arrondissement"},"description":"Ex: 75011"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cyclable/stats":{"get":{"tags":["🛣️ Aménagements Cyclables"],"summary":"Statistiques des aménagements cyclables","operationId":"get_stats_cyclable_api_cyclable_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}},"/":{"get":{"tags":["Health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-KEY"}}}}